+++ fabrictestbed_extensions/fablib/fablib.py
@@ -573,6 +573,10 @@ class FablibManager:
         "CRITICAL": logging.CRITICAL,
     }
 
+    DEFAULT_FABRIC_CREDMGR_HOST = "cm.fabric-testbed.net"
+    DEFAULT_FABRIC_ORCHESTRATOR_HOST = "orchestrator.fabric-testbed.net"
+    DEFAULT_FABRIC_BASTION_HOST = "bastion.fabric-testbed.net"
+
     default_fabric_rc = os.environ["HOME"] + "/work/fabric_config/fabric_rc"
     default_log_level = "DEBUG"
     default_log_file = "/tmp/fablib/fablib.log"
@@ -783,6 +787,15 @@ class FablibManager:
 
         self.bastion_private_ipv4_addr = "0.0.0.0"
         self.bastion_private_ipv6_addr = "0:0:0:0:0:0"
+
+        if self.credmgr_host is None:
+            self.credmgr_host = self.DEFAULT_FABRIC_CREDMGR_HOST
+
+        if self.orchestrator_host is None:
+            self.orchestrator_host = self.DEFAULT_FABRIC_ORCHESTRATOR_HOST
+
+        if self.bastion_public_addr is None:
+            self.bastion_public_addr = self.DEFAULT_FABRIC_BASTION_HOST
 
         self._validate_configuration()
 
@@ -1345,8 +1358,6 @@ class FablibManager:
         :param avoid: list of site names to avoid chosing
         :type site_name: List[String]
         :param unique:
-        :param latlon: convert address to latlon if needed for the filter, False by default
-        :type latlon: bool
         :return: list of random site names.
         :rtype: List[Sting]
         """
@@ -1685,6 +1696,7 @@ class FablibManager:
                 username=bastion_username,
                 key_filename=bastion_key_path,
                 passphrase=bastion_key_passphrase,
+                allow_agent=False,
                 look_for_keys=False,
             )
 
@@ -1693,12 +1705,14 @@ class FablibManager:
                 logging.info(f"Connection with {bastion_host} appears to be working")
                 return True
 
-        # In theory paramiko can raise several types of exceptions,
-        # but in practice it has not been that precise.  Let us treat
-        # all exceptions as un-recoverable for now, and refine the
-        # behavior based on some real-world testing.
-        except (paramiko.SSHException, Exception) as e:
-            logging.error(f"Bastion connection error: {e}")
+        except paramiko.SSHException as e:
+            logging.error(
+                f"Error connecting to bastion host {bastion_host} "
+                f"(hint: check your bastion key setup?): {e}"
+            )
+            raise e
+        except Exception as e:
+            logging.error(f"Error connecting to bastion host {bastion_host}: {e}")
             raise e
 
         finally: