+++ tests/integration/test_bastion_ssh.py
@@ -26,9 +26,7 @@ class BastionHostTests(unittest.TestCase):
         Test bastion with an empty username.
         """
         fm = FablibManager(offline=True, bastion_username="")
-        self.assertRaises(
-            paramiko.ssh_exception.AuthenticationException, fm.probe_bastion_host
-        )
+        self.assertRaises(paramiko.ssh_exception.SSHException, fm.probe_bastion_host)
 
     def test_probe_bastion_host_empty_key(self):
         """
@@ -37,9 +35,7 @@ class BastionHostTests(unittest.TestCase):
         keyfile = tempfile.NamedTemporaryFile()
 
         fm = FablibManager(offline=True, bastion_key_filename=keyfile.name)
-        self.assertRaises(
-            paramiko.ssh_exception.AuthenticationException, fm.probe_bastion_host
-        )
+        self.assertRaises(paramiko.ssh_exception.SSHException, fm.probe_bastion_host)
 
     def test_probe_bastion_host_bad_key(self):
         """
@@ -51,6 +47,4 @@ class BastionHostTests(unittest.TestCase):
         rsa_key.write_private_key_file(keyfile.name)
 
         fm = FablibManager(offline=True, bastion_key_filename=keyfile.name)
-        self.assertRaises(
-            paramiko.ssh_exception.AuthenticationException, fm.probe_bastion_host
-        )
+        self.assertRaises(paramiko.ssh_exception.SSHException, fm.probe_bastion_host)